Move income-related elements to the 'income' element group folder
