+++ angular/components/elements/element-group-income/income-input-field/models/IncomeInputContent.ts
