+++ angular/components/elements/element-group-income/income-input-fields/income-input-fields.component.scss
@@ -1,0 +1,10 @@
+// income-input-fields.component.scss
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
+
+income-input-fields {
+    @include flex(center, center, column);
+    width: 100%;
+}
