+++ angular/components/elements/element-group-income/income-input/income-input.component.scss
@@ -1,8 +1,8 @@
 // income-input.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
 
 income-input {
     position: relative;
