+++ angular/components/elements/element-group-income/income/models/types.ts
@@ -1,4 +1,4 @@
 import { EventEmitter } from '@angular/core';
-import { OneTimeIncome, RecurringIncome, WageIncome } from '../../../../models/paycheck/types';
+import { OneTimeIncome, RecurringIncome, WageIncome } from '../../../../../models/paycheck/types';
 
 export type IncomeEmitter = EventEmitter<OneTimeIncome | RecurringIncome | WageIncome>;
