+++ angular/components/elements/element-group-income/recurring-income-dropdown/models/RecurringIncomeTimeInterval.ts
