+++ angular/components/elements/element-group-income/recurring-income-dropdown/recurring-income-dropdown.component.scss
@@ -1,0 +1,11 @@
+// recurring-income-dropdown.component.scss
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
+
+recurring-income-dropdown {
+    @include flex(center, center, row);
+    width: calc($fourteen-percent * 0.935);
+    height: $three-percent;
+}
