--- angular/components/elements/income-input-field/income-input-field.component.scss
@@ -1,10 +1,0 @@
-// income-input-field.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
-@import "../element-group-bill/bill-input-field/bill-input-field.component";
-
-income-input-field {
-    @extend bill-input-field;
-}
