--- angular/components/elements/income-input-fields/income-input-fields.component.scss
@@ -1,10 +1,0 @@
-// income-input-fields.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
-
-income-input-fields {
-    @include flex(center, center, column);
-    width: 100%;
-}
